requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
MarkupSafe==1.1.1
numpy==1.23.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.9.3
six==1.16.0
Werkzeug==1.0.1
h5py==3.7.0
tensorflow==2.10.0
pillow==9.3.0
Keras==2.10.0
